Record users' certificate serial numbers.
In order to eventually disable users based on the CRLs, ATAT needs to
record users' certificate serial numbers. We record whatever serial
number is presented when they log in. If they present a different
certificate on a subsequent login, we record the new serial number. This
is based around two assumptions validated with CCPO:

- DISA will never unrevoke revoked certificate serial numbers.
- When a user's certificate it revoked, DISA will revoke all their
  certificates, so all of their certificate serial numbers will be
  published to CRLs. This means recording any one of their serial
  numbers is sufficient.