Additional validation and escaping for file names.
This adds additional front and backend validations for task order file
names. Additionally, it escapes HTML characters in file names anywhere
they are used on the frontend. We are now restricting file names to a
whitelist regex of [A-Za-z0-9\-_ \.] for simplicity.

Note:
On the frontend, the filename string must have at least one character.
This is not true in the backend validation; because of the way the
entire task order form is validated, requiring input would break the
business logic currently implemented.