Store issuer hash for user's certificates.
Storing the serial number alone for a user's certificate is not enough
to positively recognize a cert. A serial number is only unique with
reference to a given CA. We need to also store the issuer hash so that
we can correctly identify revoked users based on their serial number and
the hash of the issuer.

I chose to implement this by making the parsed certificate available
from the CRLCache. This is slightly awkward but is the path of least
resistance for getting that bit of information.